Umount for apps on denylist
NeoZygisk updates its status by mounting the `module.prop` file, which can be easily checked by app aprocess.
In the implementation, we rather directly call revert_unmount_* functions, which unmount more than the aforementioned `module.prop` file.
Consequently, the so-called Enforce DenyList function is always on for processes on denylist.

Moreover, reading of the file `/proc/self/mountinfo` can be detected by the target process, so we cache it after Zygote is started.

We sort the umounting targets to avoid errors.