Fix hardened_malloc error
DivestOS users reported that vector::resize might trigger `hardened_malloc fatal error: detected write after free`.
From the error message, this can only happen when the for loop is executed for multiple times.
Note that the for loop is effective only when data to read still exist in the socket queue, which is the case when we call `recv` with flag MSG_PEEK.
Hence, after we call `recv` with zero flags, which consume data in the socket queue, we should break the for loop.
Note: previously, we break the for loop according to errno == EAGAIN.

As a simplification, we also move the resize call to the initialization of buf.