Avoid racing issue of creating clean mount profile
There are two calls of `update_mount_namespace` with parameter `zygiskd::MountNamespace::Clean`.
In some rare case, this causes a racing issue, for example, between the daemon requests of [com.android.systemui] and [webview_zygote].

Hence, when `systemui` is started, we should make a clean mount profile for it as soon as possible.
For this reason, we postpone the caching of mount profiles after sending back the file descriptor.
This is not a perfect solution, since we haven't ensured that a clean mount profile can be created before `update_mount_namespace` is called for the second time.