Fix traces left in proc_pid_maps (#36)
Previously, we were using a pre-release version of NDK (r29), which causes `munmap` fail to completely unload `libzygisk.so`.
Returning back to the LTS version (r27) fixes this bug.

Moreover, we split the code of spoofing virtual maps into an independent function, and postone calls to it, if needed, to the moment before we unload `libzygisk.so`.