Clean KernelSU trace from the stack of Zygote
KernelSU leaves a structured data trace on the stack of the Zygote process. This trace, which contains mount information and a "KSU" identifier, is a well-known detection vector as it is inherited by all forked applications, including the System Server.

We introduce a cleaning routine that runs just before the System Server is forked (`nativeForkSystemServer_pre`). It operates by:

1.  Identifying the main thread's stack (`[anon:stack_and_tls:main]`).
2.  Searching for the unique KSU data structure's byte pattern.
3.  Intelligently calculating the full size of the artifact by skipping through its null-terminated fields.
4.  Overwriting the entire memory block with zeros using `memset`.

By cleaning the Zygote stack at this critical moment, the System Server and all subsequent applications are forked from a clean state, effectively neutralizing this detection vector.

Note that the System Server is not considered as an application process, we thus skip setting the IS_FIRST_PROCESS flag for it in the daemon.