Force recompaction of g_array to remove atexit fingerprint (#50)
The unloading of Zygisk modules can leave a persistent artifact in the internal state of libc's global `AtexitArray` (`g_array`). Specifically, the `extracted_count_` member is left non-zero, creating a detectable fingerprint.

This occurs because the standard `recompact()` logic within `libc` is not always triggered upon module unload.

This commit introduces a mechanism to manually clean this artifact after all modules are unloaded. A proxy class is used to interface with the live `g_array` object, and a custom `recompact()` function is called to unconditionally restore its state to that of a pristine process, completely removing the detection vector.