Fix regression of the `zygiskd` refactor (#57)
The condition for caching a clean mount namespace was modified by mistake.

In particular, without checking mounts whose root paths start with `/adb/modules`, NeoZygisk fails to clean the `dex2oat` mount of LSPosed. To reproduce this issue reliably, one should try JingMatrix/LSPosed#429. I wasn't able to spot this regression with the master branch of JingMatrix/LSPosed, which was the cause of this regression.