Refine rules for skipping Zygote unmount
This commit enhances the logic for determining when to skip the unmounting of modules from the Zygote process to support more complex module scenarios.

A new rule has been added to prevent unmounting when a module mounts an APK into `/system/priv-app`. This is necessary to allow modules that perform system-wide application installations to function correctly.

Additionally, the custom `starts_with` function is replaced with the standard C++ `std::string::starts_with` for improved code clarity and maintainability.